refactor(storage): move interface to storage.ts
