+++ project mermaid.txt
@@ -24,7 +24,7 @@ TikTokAffiliator/
 
 Key Files:
 1. Bot Configuration: server/bot/tiktokBot.ts
-2. Storage Interface: server/storage/index.ts
+2. Storage Interface: server/storage.ts
 3. API Routes: server/routes/botRoutes.ts
 4. Type Definitions: shared/schema.ts
 5. Tests: test/bot/*
